feat(billing-page): revert back aws support
