+++ src/pages/panel/workspace-settings-billing/utils/index.ts
@@ -1,4 +1,5 @@
 export { paymentMethodToLabel } from './paymentMethodToLabel'
+export { paymentMethods } from './paymentMethods'
 export { productTierToDescription } from './productTierToDescription'
 export { productTierToLabel } from './productTierToLabel'
 export { useGetProductTierFromSearchParams } from './useGetProductTierFromSearchParams'
