+++ src/pages/panel/workspace-settings-billing/utils/paymentMethods.ts
@@ -1,0 +1,3 @@
+import { PaymentMethods } from 'src/shared/types/server'
+
+export const paymentMethods: PaymentMethods[] = ['aws_marketplace', 'stripe']
