+++ src/shared/link-button/LinkButton.tsx
@@ -1,11 +1,11 @@
 import OpenInNewIcon from '@mui/icons-material/OpenInNew'
-import { LoadingButton } from '@mui/lab'
-import { ButtonProps, ButtonTypeMap } from '@mui/material'
+import { LoadingButton, LoadingButtonProps } from '@mui/lab'
+import { ButtonTypeMap } from '@mui/material'
 import { ElementType, useState } from 'react'
 
 // eslint-disable-next-line @typescript-eslint/ban-types
 export function LinkButton<RootComponent extends ElementType = ButtonTypeMap['defaultComponent'], AdditionalProps = {}>(
-  props: ButtonProps<RootComponent, AdditionalProps>,
+  props: LoadingButtonProps<RootComponent, AdditionalProps>,
 ) {
   const [loading, setLoading] = useState(false)
