+++ src/shared/types/server/requests/PutWorkspaceBilling.ts
@@ -1,5 +1,6 @@
-import { ProductTier } from 'src/shared/types/server'
+import { PaymentMethod, ProductTier } from 'src/shared/types/server'
 
 export interface PutWorkspaceBillingRequest {
-  product_tier: ProductTier
+  product_tier: ProductTier | null
+  workspace_payment_method: PaymentMethod | null
 }
