+++ src/shared/types/server/shared/Payment.ts
@@ -1,5 +1,7 @@
 export type PaymentMethods = 'aws_marketplace' | 'stripe' | 'none'
 
-export type PaymentMethod = { method: Exclude<PaymentMethods, 'none'>; subscription_id: string } | { method: 'none' }
+export type PaymentMethodWithoutNone = { method: Exclude<PaymentMethods, 'none'>; subscription_id: string }
+
+export type PaymentMethod = PaymentMethodWithoutNone | { method: 'none' }
 
 export type ProductTier = 'Free' | 'Trial' | 'Plus' | 'Business' | 'Enterprise'
