+++ src/shared/types/server/shared/index.ts
@@ -1,2 +1,2 @@
 export type { NotificationChannel } from './NotificationChannel'
-export type { PaymentMethod, PaymentMethods, ProductTier } from './Payment'
+export type { PaymentMethod, PaymentMethodWithoutNone, PaymentMethods, ProductTier } from './Payment'
